Split out PauliSum into separate crate
