+++ crates/ppvm-paulisum/Cargo.toml
@@ -2,3 +2,49 @@
 name = "ppvm-paulisum"
 version = "0.1.0"
 edition = "2024"
+
+[dependencies]
+ahash = { version = "0.8.12", optional = true }
+approx = { version = "0.5.1", optional = true }
+bitvec = "1.0.1"
+bon = "3.7.2"
+dashmap = { version = "6.1.0", features = ["rayon"], optional = true }
+fxhash = "0.2.1"
+gxhash = { version = "3.5.0", features = ["deterministic"], optional = true }
+indexmap = { version = "2.11.4", optional = true }
+itertools = "0.14.0"
+num = "0.4.3"
+ppvm-runtime = { version = "0.1.0", path = "../ppvm-runtime" }
+rayon = { version = "1.11.0", optional = true }
+
+[dev-dependencies]
+criterion = "0.7.0"
+insta = { version = "1.43.2", features = ["yaml"] }
+
+[features]
+default = ["dashmap", "gxhash", "ahash", "indexmap", "approx"]
+ahash = ["dep:ahash"]
+approx = ["dep:approx"]
+dashmap = ["dep:dashmap", "dep:rayon"]
+gxhash = ["dep:gxhash"]
+indexmap = ["dep:indexmap"]
+
+[[bench]]
+name = "gate"
+harness = false
+
+[[bench]]
+name = "random-circuit"
+harness = false
+
+[[bench]]
+name = "trotter"
+harness = false
+
+[[bench]]
+name = "trotter-scaling"
+harness = false
+
+[[bench]]
+name = "micro"
+harness = false
