+++ crates/ppvm-paulisum/benches/trotter-scaling.rs
