+++ crates/ppvm-paulisum/benches/trotter.rs
