+++ crates/ppvm-paulisum/src/config/gxhash.rs
@@ -4,8 +4,9 @@
 use std::collections::HashMap;
 use std::marker::PhantomData;
 
-use crate::traits::{Coefficient, NoStrategy, PauliWordTrait, Strategy};
-use crate::{config::Config, word::PauliWord};
+use ppvm_runtime::traits::{Coefficient, NoStrategy, PauliWordTrait, Strategy};
+use ppvm_runtime::config::Config;
+use ppvm_runtime::word::PauliWord;
 
 /// `HashMap`-backed [`Config`] with `[u8; N]` storage and `gxhash`.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
