+++ crates/ppvm-paulisum/src/sum/ops.rs
@@ -5,10 +5,11 @@ use std::ops::{AddAssign, MulAssign};
 
 use num::One;
 
-use crate::phase::PhasedPauliWord;
-use crate::traits::*;
-use crate::word::PauliWord;
-use crate::{config::Config, sum::PauliSum};
+use ppvm_runtime::phase::PhasedPauliWord;
+use ppvm_runtime::traits::*;
+use ppvm_runtime::word::PauliWord;
+use ppvm_runtime::config::Config;
+use crate::sum::PauliSum;
 
 /// Implement multiplication by a scalar coefficient for PauliSum.
 /// Use this macro to implement `MulAssign` for other scalar types
