+++ crates/ppvm-paulisum/src/sum/rot2.rs
@@ -2,8 +2,10 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use super::rot1::rotate_1_map_insert_closure;
-use crate::traits::*;
-use crate::{char::Pauli, config::Config, sum::PauliSum};
+use ppvm_runtime::traits::*;
+use ppvm_runtime::char::Pauli;
+use ppvm_runtime::config::Config;
+use crate::sum::PauliSum;
 
 const PAULIS: [Pauli; 4] = [Pauli::I, Pauli::X, Pauli::Z, Pauli::Y];
